+++ alodataset/__init__.py
@@ -3,7 +3,7 @@ from .sequence_mixin import SequenceMixin
 from .split_mixin import SplitMixin
 from .waymo_dataset import WaymoDataset
 from .coco_detection_dataset import CocoDetectionDataset
-from .coco_segmentation_dataset import CocoSegementationDataset
+from .coco_panoptic_dataset import CocoPanopticDataset
 from .chairssdhom_dataset import ChairsSDHomDataset
 from .flyingthings3D_subset_dataset import FlyingThings3DSubsetDataset
 from .mot17 import Mot17
