+++ alonet/detr/__init__.py
@@ -1,3 +1,7 @@
+from .data_modules import Data2Detr
+from .data_modules import CocoDetection2Detr
+from .data_modules import CocoPanoptic2Detr
+
 from .matcher import DetrHungarianMatcher
 from .criterion import DetrCriterion
 
@@ -5,8 +9,6 @@ from .detr import Detr
 from .detr_r50 import DetrR50
 from .detr_r50_finetune import DetrR50Finetune
 
-
-from .coco_data_modules import CocoDetection2Detr
 from .train import LitDetr
 from .callbacks import DetrObjectDetectorCallback
