+++ alonet/detr/data_modules/__init__.py
@@ -1,0 +1,3 @@
+from .data2detr import Data2Detr
+from .coco_detection2detr import CocoDetection2Detr
+from .coco_panoptic2detr import CocoPanoptic2Detr
